Look at README files as well.
Some modules only have the license in the README, and sometimes
when installing packages via npm the "readme" field in package.json
fails to get the README contents (instead it says "ERROR: No README data
found!" in package.json). There is certainly an npm bug somewhere, but
it seems reasonable for license-checker to directly examine the README
as well.

I tweaked the logic around the direct file checks to not overwrite the license
info with the results of each file. For example, if LICENSE has the license and
the README doesn't, it should always prefer the result it got from reading the
LICENSE>